refactor: move to egoist.app
